@@ -11,46 +11,45 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^9.0.0",
-    "@angular/cdk": "^9.0.0",
-    "@angular/common": "^9.0.0",
-    "@angular/compiler": "^9.0.0",
-    "@angular/core": "^9.0.0",
-    "@angular/forms": "^9.0.0",
+    "@angular/animations": "^9.1.0",
+    "@angular/cdk": "^9.2.0",
+    "@angular/common": "^9.1.0",
+    "@angular/compiler": "^9.1.0",
+    "@angular/core": "^9.1.0",
+    "@angular/forms": "^9.1.0",
     "@angular/http": "^7.2.16",
-    "@angular/material": "^9.0.0",
-    "@angular/material-moment-adapter": "^9.0.0",
-    "@angular/platform-browser": "^9.0.0",
-    "@angular/platform-browser-dynamic": "^9.0.0",
-    "@angular/router": "^9.0.0",
-    "@angular/service-worker": "^9.0.0",
-    "@manekinekko/angular-web-bluetooth": "^7.0.0",
+    "@angular/material": "^9.2.0",
+    "@angular/material-moment-adapter": "^9.2.0",
+    "@angular/platform-browser": "^9.1.0",
+    "@angular/platform-browser-dynamic": "^9.1.0",
+    "@angular/router": "^9.1.0",
+    "@angular/service-worker": "^9.1.0",
     "core-js": "^3.6.4",
     "material-design-icons": "^3.0.1",
     "moment": "^2.24.0",
     "ngx-toastr": "^10.2.0",
-    "rxjs": "^6.5.4",
-    "tslib": "^1.10.0",
+    "rxjs": "^6.5.5",
+    "tslib": "^1.11.1",
     "typescript": "3.7.5",
-    "zone.js": "~0.10.2"
+    "zone.js": "^0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.900.1",
-    "@angular/cli": "^9.0.1",
-    "@angular/compiler-cli": "^9.0.0",
-    "@angular/language-service": "^9.0.0",
-    "@types/jasmine": "^3.5.3",
+    "@angular-devkit/build-angular": "^0.900.7",
+    "@angular/cli": "^9.1.0",
+    "@angular/compiler-cli": "^9.1.0",
+    "@angular/language-service": "^9.1.0",
+    "@types/jasmine": "^3.5.10",
     "@types/jasminewd2": "^2.0.8",
-    "@types/node": "^12.11.1",
+    "@types/node": "^12.12.34",
     "@types/web-bluetooth": "0.0.4",
-    "codelyzer": "^5.1.2",
+    "codelyzer": "^5.2.2",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^4.4.1",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "^2.1.1",
     "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.5.2",
+    "karma-jasmine-html-reporter": "^1.5.3",
     "protractor": "^5.4.3",
     "ts-node": "~8.2.0",
     "tslint": "~5.17.0"
